api_fastapi/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pydantic
python-dotenv
google-generativeai
celery
redis
pytest
httpx
sqlalchemy
psycopg2-binary
alembic
firebase-admin